requirements.txt
google-cloud-aiplatform
streamlit
python-dotenv
altair